requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
cryptography==47.0.0
deprecation==2.1.0
fastapi==0.136.1
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
idna==3.13
multidict==6.7.1
oauthlib==3.3.1
packaging==26.2
postgrest==2.29.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==7.34.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
PyJWT==2.12.1
pyparsing==3.3.2
python-multipart==0.0.27
python-telegram-bot==22.7
realtime==2.29.0
requests==2.33.1
requests-oauthlib==2.0.0
starlette==1.0.0
supabase-auth==2.29.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.46.0
websockets==15.0.1
yarl==1.23.0